Reapply commit ab278b609cfe1fee9704abb563679ae952bcc47f

+++ test.utils/src/main/java/org/apache/clerezza/test/utils/RandomGraph.java
@@ -172,7 +172,7 @@ public class RandomGraph extends GraphWrapper {
     }
 
     private static int rollDice(int faces) {
-        return Double.valueOf(Math.random() * faces).intValue();
+        return (int) (Math.random() * faces);
     }
 
     private RDFTerm createRandomRDFTerm() {
